Revert "Chore/Move mock data to root directory (#304)"
