+++ src/services/os-places/os-places-stub.js
@@ -1,4 +1,4 @@
-import { mockAddresses } from '../../../mock-data/os-places-mock-addresses.js'
+import { mockAddresses } from '../../mock-data/mock-os-places-addresses.js'
 
 /**
  * Retrieves mock address data matching a given postcode.
