+++ test/unit/services/os-places/os-places-stub.test.js
@@ -5,7 +5,7 @@ import { describe, test, expect } from 'vitest'
 import { mockPostcode } from '../../../../src/services/os-places/os-places-stub.js'
 
 // Test helpers
-import { mockAddresses } from '../../../../mock-data/os-places-mock-addresses.js'
+import { mockAddresses } from '../../../../src/mock-data/mock-os-places-addresses.js'
 
 describe('mockPostcode', () => {
   test('returns matching addresses for a valid postcode (ignoring case and spaces)', () => {
